Added OpenLayers to the project.  Before releasing we can pick apart what we're using and what we're not.
Added functionality to handle rotation of the map.  Created a class called OpenLayers.RotatingEvents, which inherits from OpenLayers.Events, and modifies the position of the mouse to handle the rotation.  I modified Map.js within OpenLayers to use this class instead of OpenLayers.Events.

The jquery-animate-css-rotate-scale.js and jquery-css-transform.js are being used to rotate the map div.

I also added an OpenLayers.ZoomPanel to the bottom of the map.