If the tile images are not found, a white box will be displayed instead.  The nothing.png is a totally transparent image that gets set to the source so that the broken image icon disappears.  Then the background color of that img is set to white.
On initialization, the map will be created.  If geolocation is successful, draw the blue arrow at the clients location and zoom to it.  If it's not, draw the blue arrow in dc and zoom to that.  Once geolocation is successful, the arrow will be moved to the proper location and the client will be brought there.

The map does rotate according to the compass headings, but the headings returned by the navigator.compass.watchHeading() function are not very accurate.  Something to look into in the future.